requirements/base.txt
celery==5.4.0
diskcache==5.6.3
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.7.0
django-clickhouse-backend==1.3.1
django-db-connection-pool==1.2.5
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.2
mysqlclient==2.2.6
pyotp==2.9.0
user-agents==2.2.0
loguru==0.7.3
IP2Location==8.10.4
lz4==4.3.3
